Add existing packages into a package duke
