+++ src/main/java/duke/data/exception/DukeException.java
@@ -1,4 +1,4 @@
-package data.exception;
+package duke.data.exception;
 
 public class DukeException extends Exception {
     public DukeException(String message) {
